requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2.4
django-widget-tweaks==1.5.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
yarl==1.20.1
